# openapi.yaml
openapi: 3.1.0
info:
  title: Dental Lab CRM API
  version: 0.2.0-mock
  summary: API-only mock contract for web and future mobile clients
  description: |
    Draft Swagger/OpenAPI specification for `dental-lab-crm`.
    The current runtime is API-only demo mode: auth is bypassed, Prisma and
    Supabase are disconnected, and mutations are in-memory for the active
    server process.

    Reference entities used by cases are archived instead of hard-deleted.
    Archived references are hidden from default management lists and creation
    option/dropdown endpoints, but historical case reads still include related
    archived records.
servers:
  - url: http://localhost:3000
    description: Local development
security: []
...
paths:
...
  /api/cases:
    get:
      tags: [Cases]
      summary: List cases
      description: Returns lightweight case summaries inside the standard response envelope. Included case reference names are returned even when the referenced row has since been archived. Rich case detail, processes, dependencies, attachments, components, and milling rows are available from GET /api/cases/{id}.
      operationId: listCases
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 100
          description: Maximum number of newest matching cases to return.
        - in: query
          name: status
          schema:
            $ref: "#/components/schemas/CaseStatus"
        - in: query
          name: customerId
          schema:
            type: string
        - in: query
          name: urgent
          schema:
            type: string
            enum: [urgent, normal, "true", "false"]
        - in: query
          name: q
          schema:
            type: string
          description: Search by case code, client case code, patient, customer, dentist, service type, or CAD designer.
        - in: query
          name: search
          schema:
            type: string
          deprecated: true
          description: Alias for q.
      responses:
        "200":
          description: Case list returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseListResponse"
    post:
      tags: [Cases]
      summary: Create case
      description: Creates a new dental case entry for the authenticated user's dental lab. Case codes are generated by the backend from existing numeric codes in that lab; any request body `code` value is ignored. Selected reference ids must be active, unarchived, and scoped to the same lab.
      operationId: createCase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseRequest"
      responses:
        "201":
          description: Case created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseMutationResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/cases/{id}:
    get:
      tags: [Cases]
      summary: Get case details
      description: Returns the full detail of a single case from the in-memory mock store.
      operationId: getCaseById
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Case detail returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseResponse"
        "404":
          description: Case not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags: [Cases]
      summary: Update case
      description: Updates editable case fields in the in-memory mock store.
      operationId: updateCase
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCaseRequest"
      responses:
        "200":
          description: Case updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseMutationResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Cases]
      summary: Delete case
      description: Deletes a case and related mock production records from the in-memory mock store.
      operationId: deleteCase
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Case deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteEntityResponse"
        "400":
          description: Case cannot be deleted because attachments are still retained
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Case not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/case-processes/{id}:
    patch:
      tags: [Production]
      summary: Update case process
      description: Updates a real case workflow step. When status becomes completed, dependent locked steps are marked ready after all dependencies are completed.
      operationId: updateCaseProcess
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCaseProcessRequest"
      responses:
        "200":
          description: Case process updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseProcessMutationResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Case process not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/customers/{id}:
    patch:
      tags: [Registry]
      summary: Update customer
      description: Updates a customer by id.
      operationId: updateCustomer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistryEntityPayload"
      responses:
        "200":
          description: customer updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistryEntityMutationResponse"
        "404":
          description: Registry entity not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags: [Registry]
      summary: Archive customer
      description: Archives a customer by setting `isActive=false` and `deletedAt` to the current timestamp.
      operationId: archiveCustomer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: customer archived successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveEntityResponse"
  /api/cad-designers/{id}:
    patch:
      tags: [Registry]
      summary: Update CAD designer
      description: Updates a CAD designer by id.
      operationId: updateCadDesigner
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistryEntityPayload"
      responses:
        "200":
          description: CAD designer updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistryEntityMutationResponse"
    delete:
      tags: [Registry]
      summary: Archive CAD designer
      description: Archives the CAD designer user row by setting `isActive=false` and `deletedAt` to the current timestamp.
      operationId: archiveCadDesigner
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: CAD designer archived successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveEntityResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Explain what went wrong
    ErrorResponse:
      type: object
      required: [data, error, meta]
      properties:
        data:
          type: "null"
          example: null
        error:
          $ref: "#/components/schemas/ErrorObject"
        meta:
          type: object
          required: [limit]
          properties:
            limit:
              type: integer
              example: 100
...
    CaseStatus:
      type: string
      enum:
        [
          ENTRY,
          WAITING_INFO,
          DESIGNING,
          WAITING_APPROVAL,
          DESIGN_READY,
          MILLING_PRINTING,
          DONE,
        ]
    CaseProcessStatus:
      type: string
      enum: [locked, ready, in_progress, completed, skipped, cancelled]
    AttachmentKind:
      type: string
      enum: [SCAN_INPUT, DESIGN_OUTPUT, MODEL_OUTPUT, OTHER]
...
    CaseProcess:
      type: object
      required: [id, processId, workflowStepId, status, createdAt, updatedAt]
      properties:
        id:
          type: string
          example: case_process_001
        processId:
          type: string
          example: process_001
        processName:
          type: string
          example: Design
        workflowStepId:
          type: string
          example: design
        status:
          $ref: "#/components/schemas/CaseProcessStatus"
        assignedToId:
          type: string
          nullable: true
        assignedToName:
          type: string
          nullable: true
        dependsOnCaseProcessIds:
          type: array
          items:
            type: string
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CaseAttachment:
      type: object
      required: [id, kind, fileName, uploadedAt]
      properties:
        id:
          type: string
          example: attach_001
        kind:
          $ref: "#/components/schemas/AttachmentKind"
        fileName:
          type: string
          example: scan-upper.stl
        uploadedAt:
          type: string
          format: date-time
    CaseTimelineItem:
      type: object
      required: [id, toStatus, changedAt]
      properties:
        id:
          type: string
          example: hist_001
        fromStatus:
          allOf:
            - $ref: "#/components/schemas/CaseStatus"
          nullable: true
        toStatus:
          $ref: "#/components/schemas/CaseStatus"
        note:
          type: string
          nullable: true
          example: CAD aprovado
        changedAt:
          type: string
          format: date-time
    CaseSummary:
      type: object
      required:
        [id, code, patientName, currentStatus, isUrgent, createdAt, updatedAt]
      properties:
        id:
          type: string
          example: case_001
        dentalLabId:
          type: string
        code:
          type: string
          example: "0587"
        clientCaseCode:
          type: string
          nullable: true
          example: CL-44A
        patientName:
          type: string
          example: João Silva
        customerId:
          type: string
          nullable: true
        customerName:
          type: string
          nullable: true
          example: Smile Center
        dentistId:
          type: string
          nullable: true
        dentistName:
          type: string
          nullable: true
          example: Dra. Ana
        serviceTypeId:
          type: string
          nullable: true
        serviceTypeName:
          type: string
          nullable: true
          example: Crown
        cadDesignerId:
          type: string
          nullable: true
        cadDesignerName:
          type: string
          nullable: true
          example: Camila Santos
        createdByUserId:
          type: string
          nullable: true
        createdByUserName:
          type: string
          nullable: true
          example: Roberto Lima
        currentStatus:
          $ref: "#/components/schemas/CaseStatus"
        dueDate:
          type: string
          format: date-time
          nullable: true
        isUrgent:
          type: boolean
          example: false
        teeth:
          type: string
          nullable: true
          example: 11,12
        elementsQty:
          type: integer
          nullable: true
          example: 2
        shade:
          type: string
          nullable: true
          example: A2
        observations:
          type: string
          nullable: true
        pendingNote:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CaseDetail:
      allOf:
        - $ref: "#/components/schemas/CaseSummary"
        - type: object
          properties:
            processes:
              type: array
              items:
                $ref: "#/components/schemas/CaseProcess"
            attachments:
              type: array
              items:
                $ref: "#/components/schemas/CaseAttachment"
            timeline:
              type: array
              items:
                $ref: "#/components/schemas/CaseTimelineItem"
...
    CreateCaseRequest:
      type: object
      required: [patientName]
      properties:
        customerId:
          type: string
          nullable: true
          example: customer_001
        serviceTypeId:
          type: string
          nullable: true
          example: service_001
        dentistId:
          type: string
          nullable: true
          example: dentist_001
        cadDesignerId:
          type: string
          nullable: true
          example: 8f5ee9f0-8d21-4ad7-b50b-0123456789ab
        patientName:
          type: string
          example: João Silva
        dueDate:
          type: string
          format: date
          nullable: true
        isUrgent:
          type: boolean
          default: false
        teeth:
          type: string
          nullable: true
        elementsQty:
          type: integer
          nullable: true
        shade:
          type: string
          nullable: true
        observations:
          type: string
          nullable: true
    UpdateCaseRequest:
      type: object
      properties:
        patientName:
          type: string
        customerId:
          type: string
          nullable: true
        serviceTypeId:
          type: string
          nullable: true
        dentistId:
          type: string
          nullable: true
        cadDesignerId:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date
          nullable: true
        isUrgent:
          type: boolean
        teeth:
          type: string
          nullable: true
        elementsQty:
          type: integer
          nullable: true
        shade:
          type: string
          nullable: true
        observations:
          type: string
          nullable: true
        pendingNote:
          type: string
          nullable: true
...
    CaseListResponse:
      type: object
      required: [data, error, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CaseSummary"
        error:
          nullable: true
          example: null
        meta:
          type: object
          additionalProperties: true
          example: {}
    CaseResponse:
      type: object
      required: [data, error, meta]
      properties:
        data:
          $ref: "#/components/schemas/CaseDetail"
        error:
          nullable: true
          example: null
        meta:
          type: object
          additionalProperties: true
          example: {}
    CaseMutationResponse:
      type: object
      required: [data, error, meta]
      properties:
        data:
          type: object
          properties:
            id:
              type: string
              example: case_001
            code:
              type: string
              example: "0587"
            currentStatus:
              $ref: "#/components/schemas/CaseStatus"
            updatedAt:
              type: string
              format: date-time
        error:
          nullable: true
          example: null
        meta:
          type: object
          additionalProperties: true
          example: {}
    CaseProcessMutationResponse:
      type: object
      required: [data, error, meta]
      properties:
        data:
          $ref: "#/components/schemas/CaseProcess"
        error:
          nullable: true
          example: null
        meta:
          type: object
          additionalProperties: true
          example: {}
...
    RegistryEntityPayload:
      type: object
      additionalProperties: true
      description: Generic payload. Shape depends on `entity` and is validated server-side.
...
    UpdateCaseProcessRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/CaseProcessStatus"
        assignedToId:
          type: string
          nullable: true
...
    RegistryEntityMutationResponse:
      type: object
      required: [data, error, meta]
      properties:
        data:
          type: object
          additionalProperties: true
        error:
          nullable: true
          example: null
        meta:
          type: object
          additionalProperties: true
          example: {}
    ArchiveEntityResponse:
      type: object
      required: [data, error, meta]
      properties:
        data:
          type: object
          required: [id, isActive, deletedAt]
          properties:
            id:
              type: string
            isActive:
              type: boolean
              const: false
            deletedAt:
              type: string
              format: date-time
          additionalProperties: true
        error:
          nullable: true
          example: null
        meta:
          type: object
          additionalProperties: true
          example: {}
    DeleteEntityResponse:
      type: object
      required: [data, error, meta]
      properties:
        data:
          type: object
          required: [id, deleted]
          properties:
            id:
              type: string
            deleted:
              type: boolean
              example: true
        error:
          nullable: true
          example: null
        meta:
          type: object
          additionalProperties: true
          example: {}
...